캐스팅해주기

+++ chapter14/src/sec05/exam03/FunctionExample2.java
@@ -15,7 +15,7 @@ public class FunctionExample2 {
         for(Student student : list) {
             sum += function.applyAsInt(student);
         }
-        double avg = sum / list.size();
+        double avg = (double)sum / list.size();
         return avg;
     }
